Move ErrorCode::Unknown inner value to opaque type
This ensures that adding new error types will not be a breaking
change, as unknown error codes cannot be matched on.  It is still
possible to retrieve the inner value of the unknown error code by
using the `Display` implementation or the `as_str` method, but
those are much less likely to be used and should not be expected
to be stable.